epic-4/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pytest>=7.0.0

# Optional cloud storage dependencies
google-cloud-storage>=2.10.0
boto3>=1.28.0
